Create README.md
This Calculator code was demonstrated by NeuralNine on YouTube. I followed his tutorial to follow along and build a GUI calculator using PyCharm. The tutorial was very helpful to see how a tkinter GUI is created from start to finish using Python code. I would like to add more features to this code and make some changes, but it was more important to try understand how this GUI was created.